plugins/tools-document/skills/pdf/scripts/requirements.txt
pypdf>=4.0
pdfplumber>=0.10
reportlab>=4.0
pypdfium2>=4.0
pdf2image>=1.16
pytesseract>=0.3
Pillow>=10.0
